refactor(#14): admin 디렉토리 계층 구조 도메인별로 분리
