+++ src/main/kotlin/site/billilge/api/backend/domain/item/controller/AdminItemApi.kt
@@ -1,4 +1,4 @@
-package site.billilge.api.backend.domain.admin.controller
+package site.billilge.api.backend.domain.item.controller
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
@@ -7,18 +7,15 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.responses.ApiResponses
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.*
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.RequestPart
 import org.springframework.web.multipart.MultipartFile
-import site.billilge.api.backend.domain.admin.dto.response.AdminFindAllResponse
 import site.billilge.api.backend.domain.item.dto.request.ItemRequest
 import site.billilge.api.backend.domain.item.dto.response.ItemFindAllResponse
-import site.billilge.api.backend.domain.payer.dto.request.PayerRequest
-import site.billilge.api.backend.domain.payer.dto.response.PayerFindAllResponse
 import site.billilge.api.backend.global.exception.ErrorResponse
 
-@Tag(name = "Admin", description = "관리자 API")
-interface AdminApi {
-
+@Tag(name = "(Admin) Item", description = "관리자용 물품 API")
+interface AdminItemApi {
     @Operation(
         summary = "대여 물품 목록 조회",
         description = "대여 물품 목록을 조회하는 관리자용 API"
@@ -95,69 +92,4 @@ interface AdminApi {
     fun deleteItem(
         @PathVariable itemId: Long,
     ): ResponseEntity<Void>
-
-    @Operation(
-        summary = "관리자 목록 조회",
-        description = "관리자 목록을 조회하는 API"
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "관리자 목록 조회 성공"
-            )
-        ]
-    )
-    fun getAdminList(): ResponseEntity<AdminFindAllResponse>
-
-    @Operation(
-        summary = "회원 권한 변경",
-        description = "회원의 권한(사용자/관리자)을 변경하는 관리자용 API"
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "회원이 사용자라면 관리자로, 관리자라면 사용자로 권한 변경 성공"
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "회원을 찾을 수 없습니다.",
-                content = [Content(schema = Schema(implementation = ErrorResponse::class))]
-            )
-        ]
-    )
-    fun updateMemberRole(@PathVariable memberId: Long): ResponseEntity<Void>
-
-    @Operation(
-        summary = "학생회비 납부자 추가",
-        description = "학생회비 납부자 데이터를 추가하는 관리자용 API"
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "학생회비 납부자 데이터 추가 성공"
-            )
-        ]
-    )
-    fun addPayers(@RequestBody request: PayerRequest): ResponseEntity<Void>
-
-    @Operation(
-        summary = "학생회비 납부자 목록 조회",
-        description = "학생회비 납부자 목록을 전체 혹은 페이지별로 조회하는 관리자용 API"
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "학생회비 납부자 목록 조회 성공"
-            )
-        ]
-    )
-    fun getAllPayers(
-        @RequestParam(required = false, defaultValue = "0") pageNo: Int,
-        @RequestParam(required = false, defaultValue = "10") size: Int,
-        @RequestParam(required = false, defaultValue = "enrollmentYear") criteria: String
-    ): ResponseEntity<PayerFindAllResponse>
 }
