+++ src/main/kotlin/site/billilge/api/backend/domain/member/dto/response/AdminFindAllResponse.kt
@@ -1,4 +1,4 @@
-package site.billilge.api.backend.domain.admin.dto.response
+package site.billilge.api.backend.domain.member.dto.response
 
 data class AdminFindAllResponse(
     val admins: List<AdminMemberDetail>
